% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_composite_raster.R
\name{create_composite_raster}
\alias{create_composite_raster}
\title{Create Composite raster}
\usage{
create_composite_raster(
  file1,
  file2,
  mask,
  years_input_file,
  years_raster_stack,
  dname
)
}
\arguments{
\item{file1}{file An ncdf file for mask 0 values}

\item{file2}{file An ncdf file for mask 1 values}

\item{mask}{Irrigation mask raster with values 0,1}

\item{years_input_file}{A vector indicating years in the input file}

\item{years_raster_stack}{A vector stating for which years the RasterStack should be created}

\item{dname}{A variable name. One of "Weekly Growth Index", "Hot Stress", "Dry Stress", "Wet Stress", "Cold Stress", "Ecoclimatic Index", "Moisture Index", "Temperature Index"}
}
\value{
A Composite RasterStack with some metadata
}
\description{
file1, file2, mask, years, indicatior
}
\examples{
\dontrun{
composite_raster_dorsalis <- fun1(file1 = "WeeklyGI_B,dorsalis_2018-2019_World.nc",
                                  file2 = "WeeklyGI_B,dorsalis_2018-2019_World_irrigation_set.nc",
                                  mask = "CM30_1995H_V2_gmia_v5_aei_h_classified10ha.tif",
                                  years = 2018:2019,
                                  dname = "Weekly Growth Index")

}
}
