% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_list.R
\name{extract_data_list}
\alias{extract_data_list}
\title{Extract Data List}
\usage{
extract_data_list(file, dname, n = 52, years, ...)
}
\arguments{
\item{file}{An ncdf file}

\item{dname}{A variable name. One of "Weekly Growth Index", "Hot Stress", "Dry Stress", "Wet Stress", "Cold Stress", "Ecoclimatic Index", "Moisture Index", "Temperature Index"}

\item{n}{A number of layers for each year. The default is 52}

\item{years}{A vector indicating years in the input file}

\item{...}{Not used}
}
\value{
A nested list with spatio-thematic metadata.
Each nested list contains a data for each year
in a form of a set of matrices.
}
\description{
Extract Data List
}
\examples{
\dontrun{
  Path <- "C:/Google Drive/Corvus Geostat - Jakub Nowosad/"
  Path <- "../corvus_dynamic_outputs/"
  input_file <- paste0(Path,"NetCDF/DxResults_Cc_2010-2015.nc")
  AllYears <- extract_data_list(input_file, "Weekly Growth Index",
                              years = 2000:2014)
}
}
