% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffipm-package.R
\docType{package}
\name{ffipm-package}
\alias{ffipm}
\alias{ffipm-package}
\title{ffipm: FF-IPM Support Functions}
\description{
A set of functions supporting processing and visualizing data related to the FF-IPM project. It includes creating RasterStack based on models' results, extracting time series from models, making animations, and many others.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aniaszy/ffipm}
  \item Report bugs at \url{https://github.com/aniaszy/ffipm/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

Other contributors:
\itemize{
  \item Corvus Geostat [copyright holder]
}

}
\keyword{internal}
