% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_animation.R
\name{raster_animation}
\alias{raster_animation}
\title{Raster Animation}
\usage{
raster_animation(
  rasters,
  filename,
  bbox,
  outdir,
  framerate = 4,
  width = 7,
  height,
  dpi = 100
)
}
\arguments{
\item{rasters}{Any RasterStack}

\item{filename}{The output filename with a video extension, such as mp4, mkv, mov, or flv. For more information visit \code{\link[av:encoding]{av::av_encode_video()}}}

\item{bbox}{A vector with four values corresponding to xmin, ymin, xmax, ymax}

\item{outdir}{Output directory for storing video frames (graphics). The default is a temporary folder}

\item{framerate}{A number of frames per seconds. For more information visit \code{\link[av:encoding]{av::av_encode_video()}}}

\item{width}{The width the plot. The default is 7 (inches). For more information visit \code{\link[tmap:tmap_save]{tmap::tmap_save()}}}

\item{height}{The height of the plot. The default value is calculated automatically based on the RasterStack dimensions}

\item{dpi}{The default is 100. For more information visit \code{\link[tmap:tmap_save]{tmap::tmap_save()}}}
}
\value{
A video file
}
\description{
Raster Animation
}
\examples{
 \dontrun{
  Path <- "C:/Google Drive/Corvus Geostat - Jakub Nowosad/"
  Path <- "../corvus_dynamic_outputs/"
  input_file <- paste0(Path,"NetCDF/DxResults_Cc_2010-2015.nc")
  AllYears <- extract_data_list(input_file, "Weekly Growth Index",
                              years = 2000:2014)
  rasters <- create_raster_stack(AllYears, years = 2000:2014)

  # xmin, ymin, xmax, ymax
  world_bbox <- c(-180, -90, 180, 90)
  africa_bbox <- c(-17.8, -34.9, 51.2, 37.4)
  europe_bbox <- c(-32, 34, 70, 56)
  #

  # it takes about 8 minut for 780 frames
  # the below examples use only 60 frames to show an example
  rasters2 = raster::subset(rasters, 1:60)
  raster_animation(rasters = rasters2,
                   filename = "europe6.mp4",
                   bbox = europe_bbox)

  raster_animation(rasters = rasters2,
                   filename = "africa.mp4",
                   bbox = africa_bbox)

  raster_animation(rasters = rasters2,
                   filename = "world.mp4",
                   bbox = world_bbox)
}
}
