% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pd.R
\name{read_pd}
\alias{read_pd}
\title{Read PD}
\usage{
read_pd(
  filename,
  species,
  maindir = "Digitising FF trapping plots/",
  id,
  param_df,
  ...
)
}
\arguments{
\item{filename}{Name of the file without extension as specified in the \code{File_name} column of the Diagrams_sumary - HG_zmiany.ods file}

\item{species}{"Cc", "Bd", or "Bz"}

\item{maindir}{A main directory containing subdirectories for each species}

\item{id}{Number of a row from the table/data_summary.ods file.
Can be used instead of \code{filename} and \code{species}}

\item{param_df}{A path to the file with metadata (e.g., "table/data_summary.ods")}

\item{...}{Additional arguments for \code{\link[utils:read.table]{utils::read.csv2()}}}
}
\value{
A data.frame with four columns: time, value, lat, and lon.
It additionally contains several hidden attributes: xlab, ylab, and location.
}
\description{
Read PD
}
\examples{
\dontrun{
  maindir <- "../corvus_dynamic_outputs/Digitising FF trapping plots/"
  filename <- "Bjelis_2007_1"; species <- "Cc"
  x <- read_pd(filename, species, maindir)
  y <- read_pd(maindir = maindir, id = 3)
}

}
