% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_plot.R
\name{trap_plot}
\alias{trap_plot}
\title{Trap Plot}
\usage{
trap_plot(df, type = "line")
}
\arguments{
\item{df}{The output of the \code{\link[=read_pd]{read_pd()}} function}

\item{type}{A type of the output plot. Either "line" or "bar"}
}
\value{
A ggplot object
}
\description{
Trap Plot
}
\examples{
\dontrun{
  maindir <- "../corvus_dynamic_outputs/Digitising FF trapping plots/"
  filename <- "Bjelis_2007_1"; species <- "Cc"
  x <- read_pd(filename, species, maindir)
  trap_plot(x)
  trap_plot(x, "bar")
  trap_plot(x) + ggplot2::labs(title = "Title", x = "X", y = "Y")
}
}
